+++ src/routes/FileRoute/File.route.js
@@ -1,14 +1,14 @@
 'use strict';
 
-const express = require('express');
-const bodyParser = require("body-parser");
-const fileUpload = require('express-fileupload');
-const fileSystem = require('fs');
-const path = require('path');
-const router = express.Router();
-const FileController = require("../../controllers").FileController;
-const UserController = require("../../controllers").UserController;
-const AuthController = require('../../controllers').AuthController;
+const express           = require('express');
+const bodyParser        = require("body-parser");
+const fileUpload        = require('express-fileupload');
+const fileSystem        = require('fs');
+const path              = require('path');
+const router            = express.Router();
+const FileController    = require("../../controllers").FileController;
+const UserController    = require("../../controllers").UserController;
+const AuthController    = require('../../controllers').AuthController;
 const HistoryController = require('../../controllers').HistoryController;
 
 router.use(bodyParser.json());
@@ -29,8 +29,7 @@ router.get('/', UserController.checkLevel(1), async (req, res) => {
     try {
         const files = await FileController.getAllVersions(req.params.name);
         res.json(files);
-    }
-    catch (e) {
+    } catch (e) {
         console.log(e.toString());
         res.status(409).end();
     }
@@ -47,34 +46,37 @@ router.get('/', UserController.checkLevel(1), async (req, res) => {
 
 router.post('/', async (req, res) => {
     try {
-        if(!req.files) {
+        if (!req.files) {
             console.log('No file to upload or file empty');
             return res.status(409).end();
         }
-        let g = null;
+        let g                = null;
         const isFirstVersion = await FileController.isFirstVersion(req.body.name, req.body.directory);
-        if(isFirstVersion) {
-            g = await FileController.create(req.body.name,req.body.date_create,1,req.body.file_type,req.body.user_create,req.body.user_update,req.body.directory);
-        }
-        else {
+        if (isFirstVersion) {
+            g = await FileController.create(req.body.name, req.body.date_create, 1, req.body.file_type, req.body.user_create, req.body.user_update, req.body.directory);
+        } else {
             const lastVersion = await FileController.getLastVersion(req.body.name, req.body.directory);
-            g = await FileController.create(req.body.name,req.body.date_create,parseInt(lastVersion, 10)+1,req.body.file_type,req.body.user_create,req.body.user_update,req.body.directory);
+            g                 = await FileController.create(req.body.name, req.body.date_create, parseInt(lastVersion, 10) + 1, req.body.file_type, req.body.user_create, req.body.user_update, req.body.directory);
         }
 
         let fileToUpload = req.files.file;
-        if(!fileToUpload) {
+        if (!fileToUpload) {
             console.log('Error while uploading the file');
             return res.status(409).end();
         }
 
-        fileToUpload.mv(process.env.FILES_PATH + g._id, function(err) {
+        fileToUpload.mv(process.env.FILES_PATH + g._id, function (err) {
             if (err)
                 return res.status(500).send(err);
             console.log('File uploaded!');
         });
+        console.log('parent dir :');
+        console.log(g.parent_directory);
+        HistoryController.create('created', null, g._id, null, null, req.user.id);
+        HistoryController.create('addedFile', g.directory, null, null, g._id, req.user.id);
         // const h = await HistoryController.create("upload",req.body.date_create);
         res.status(201).end();
-    } catch(err) {
+    } catch (err) {
         console.log(err.toString());
         res.status(409).end();
     }
@@ -82,30 +84,42 @@ router.post('/', async (req, res) => {
 
 router.put('/', async (req, res) => {
     const id = req.body.id;
-    if(id === undefined){
+    if (id === undefined) {
         return res.status(400).end();
     }
 
     try {
         const g = await FileController.update(id, req.body);
 
-        if(g === null || g === undefined) res.status(204).end();
-        else res.status(200).end();
-    } catch(err) {
+        if (g === null || g === undefined) res.status(204).end();
+        else {
+            HistoryController.create('modified', null, g._id, null, null, req.user.id);
+            res.status(200).end();
+        }
+    } catch (err) {
+        console.log(err);
         res.status(400).end();
     }
 });
 
 router.delete('/:id', async (req, res) => {
     const id = req.params.id;
-    if(id === undefined) {
+    if (id === undefined) {
         return res.status(400).end();
     }
     try {
-        const g = await FileController.delete(id);
-        const h = await HistoryController.create('deletion', null, null, id, null);
+        const file = await FileController.getById(id);
+
+        if(file) {
+            console.log('dir edited after delete');
+            console.log(file.directory);
+            const g = await FileController.delete(id);
+            HistoryController.create('deletedFile', file.directory, null, null, file._id, req.user.id);
+        }
+
         res.status(200).end();
-    } catch(err) {
+    } catch (err) {
+        console.log(err);
         res.status(409).end();
     }
 });
